refactor(llm): add a button to backup data & count together (#153)
1. add a button to backup data
2. refactor backup_data
3. refactor timely_update_vid_embedding
4. return count & graph elements together

---------
