+++ hugegraph-llm/.gitignore
@@ -1,2 +1,3 @@
 src/hugegraph_llm/resources/demo/questions_answers.xlsx
 src/hugegraph_llm/resources/demo/questions.xlsx
+src/hugegraph_llm/resources/backup-graph-data-4020/
